E.matrices {matrixcalc}	R Documentation

List of E Matrices

Description

This function constructs and returns a list of lists. The component of each sublist is a square matrix derived from the column vectors of an order n identity matrix.

Usage

E.matrices(n)

Arguments

n
a positive integer for the order of the identity matrix

Details

Let be the order n identity matrix with corresponding unit vectors {{{\bf{e}}_i}} with one in its ith position and zeros elsewhere. The n \times n matrix {{\bf{E}}_{i,j}} is computed from the unit vectors {{{\bf{e}}_i}} and {{{\bf{e}}_j}} as {{\bf{E}}_{i,j}} = {{\bf{e}}_i}\;{{\bf{e'}}_j}. These matrices are stored as components in a list of lists.

Value

A list with n components

1
A sublist of n components

2
A sublist of n components

...

n
A sublist of n components

Each component j of sublist i is a matrix {\bf{E}}_{i,j}

Note

The argument n must be an integer value greater than or equal to 2.

Author(s)

Frederick Novomestky fnovomes@poly.edu

References

Magnus, J. R. and H. Neudecker (1980). The elimination matrix, some lemmas and applications, SIAM Journal on Algebraic Discrete Methods, 1(4), December 1980, 422-449.

Magnus, J. R. and H. Neudecker (1999). Matrix Differential Calculus with Applications in Statistics and Econometrics, Second Edition, John Wiley.

Examples

E <- E.matrices( 3 )

[Package matrixcalc version 1.0-6 Index]