H.matrices {matrixcalc}	R Documentation

List of H Matrices

Description

This function constructs and returns a list of lists. The component of each sublist is derived from column vectors in an order r and order c identity matrix.

Usage

H.matrices(r, c = r)

Arguments

r
a positive integer value for an order r identity matrix

c
a positive integer value for an order c identify matrix

Details

Let {{\bf{I}}_r} = \lbrack {\begin{array}{cccc} {{{\bf{a}}_1}}&{{{\bf{a}}_2}}& \cdots &{{{\bf{a}}_r}} \end{array}} \rbrack be the order r identity matrix with corresponding unit vectors {{{\bf{a}}_i}} with one in its ith position and zeros elsewhere. Let {{\bf{I}}_c} = \lbrack {\begin{array}{cccc} {{{\bf{b}}_1}}&{{{\bf{b}}_2}}& \cdots &{{{\bf{b}}_c}} \end{array}} \rbrack be the order c identity matrix with corresponding unit vectors {{{\bf{b}}_i}} with one in its ith position and zeros elsewhere. The r \times c matrix {\bf{H}}{}_{i,j} = {{\bf{a}}_i}\;{{\bf{b'}}_j} is used in the computation of the commutation matrix.

Value

A list with r components

1
A sublist of c components

2
A sublist of c components

...

r
A sublist of c components

Each component j of sublist i is a matrix {\bf{H}}_{i,j}

Note

The argument n must be an integer value greater than or equal to two.

Author(s)

Frederick Novomestky fnovomes@poly.edu

References

Magnus, J. R. and H. Neudecker (1979). The commutation matrix: some properties and applications, The Annals of Statistics, 7(2), 381-394.

Magnus, J. R. and H. Neudecker (1980). The elimination matrix, some lemmas and applications, SIAM Journal on Algebraic Discrete Methods, 1(4), December 1980, 422-449.

Examples

H.2.3 <- H.matrices( 2, 3 )
H.3 <- H.matrices( 3 )

[Package matrixcalc version 1.0-6 Index]