N.matrix {matrixcalc}	R Documentation

Construct N Matrix

Description

This function returns the order n square matrix that is the sum of an implicit commutation matrix and the order n identity matrix quantity divided by two

Usage

N.matrix(n)

Arguments

n
A positive integer matrix order

Details

Let {\bf{K}_n} be the order n implicit commutation matrix (i.e., {{\bf{K}}_{n,n}} ). and {{\bf{I}}_n} the order n identity matrix. The formula for the matrix is {\bf{N}} = \frac{1}{2}≤ft( {{{\bf{K}}_n} + {{\bf{I}}_n}} \right).

Value

An order n matrix.

Note

If the argument is not an integer, the function displays an error message and stops. If the argument is less than two, the function displays an error message and stops.

Author(s)

Frederick Novomestky fnovomes@poly.edu

References

Magnus, J. R. and H. Neudecker (1980). The elimination matrix, some lemmas and applications, SIAM Journal on Algebraic Discrete Methods, 1(4), December 1980, 422-449.

Magnus, J. R. and H. Neudecker (1999) Matrix Differential Calculus with Applications in Statistics and Econometrics, Second Edition, John Wiley.

See Also

K.matrix

Examples

N <- N.matrix( 3 )
print( N )

[Package matrixcalc version 1.0-6 Index]