T.matrices {matrixcalc}	R Documentation

List of T Matrices

Description

This function constructs a list of lists. The number of components in the high level list is n. Each of the n components is also a list. Each sub-list has n components each of which is an order n square matrix.

Usage

T.matrices(n)

Arguments

n
a positive integer value for the order of the matrices

Details

Let {{\bf{E}}_{i,j}}\;i = 1, … ,n\;;\;j = 1, … ,n be a representative order n matrix created with function E.matrices. The order n matrix {{\bf{T}}_{i,j}} is defined as follows {{\bf{T}}_{i,j}} = ≤ft\{ {\begin{array}{cc} {{{\bf{E}}_{i,j}}}&{i = j}\\ {{{\bf{E}}_{i,j}} + {{\bf{E}}_{j,i}}}&{i \ne j} \end{array}} \right.

Value

A list of n components.

1
A list of n components

2
A list of n components

...

n
A list of n components

Each component j of sublist i is a matrix {\bf{T}}_{i,j}

Note

The argument n must be an integer value greater than or equal to 2.

Author(s)

Frederick Novomestky fnovomes@poly.edu

References

Magnus, J. R. and H. Neudecker (1980). The elimination matrix, some lemmas and applications, SIAM Journal on Algebraic Discrete Methods, 1(4), December 1980, 422-449.

Magnus, J. R. and H. Neudecker (1999) Matrix Differential Calculus with Applications in Statistics and Econometrics, Second Edition, John Wiley.

See Also

E.matrices

Examples

T <- T.matrices( 3 )

[Package matrixcalc version 1.0-6 Index]