creation.matrix {matrixcalc}	R Documentation

Creation Matrix

Description

This function returns the order n creation matrix, a square matrix with the sequence 1, 2, ..., n - 1 on the sub-diagonal below the principal diagonal.

Usage

creation.matrix(n)

Arguments

n
a positive integer greater than 1

Details

The order n creation matrix is also called the derivation matrix and is used in numerical mathematics and physics. It arises in the solution of linear dynamical systems. The form of the matrix is ≤ft\lbrack {\begin{array}{cccccc} 0&0&0& \cdots &0&0\\ 1&0&0& \cdots &0&0\\ 0&2&0& \cdots &0&0\\ 0&0&3& \ddots &0&0\\ \vdots & \vdots & \vdots & \ddots & \ddots &{}\\ 0&0&0& \cdots &{n - 1}&0 \end{array}} \right\rbrack.

Value

An order n matrix.

Note

If the argument n is not an integer that is greater than 1, the function presents an error message and stops.

Author(s)

Frederick Novomestky fnovomes@poly.edu

References

Aceto, L. and D. Trigiante (2001). Matrices of Pascal and Other Greats, American Mathematical Monthly, March 2001, 108(3), 232-245.

Weinberg, S. (1995). The Quantum Theory of Fields, Cambridge University Press.

Examples

H <- creation.matrix( 10 )
print( H )

[Package matrixcalc version 1.0-6 Index]