fibonacci.matrix {matrixcalc}	R Documentation

Fibonacci Matrix

Description

This function constructs the order n + 1 square Fibonacci matrix which is derived from a Fibonacci sequence.

Usage

fibonacci.matrix(n)

Arguments

n
a positive integer value

Details

Let ≤ft\{ {{f_0},\;{f_1},\; … ,\;{f_n}} \right\} be the set of n + 1 Fibonacci numbers where {f_0} = {f_1} = 1 and {f_j} = {f_{j - 1}} + {f_{j - 2}},\quad 2 ≤ j ≤ n. The order n + 1 Fibonacci matrix {\bf{F}} has as typical element {F_{i,j}} = ≤ft\{ {\begin{array}{cc} {{f_{i - j + 1}}}&{i - j + 1 ≥ 0}\\ 0&{i - j + 1 < 0} \end{array}} \right..

Value

An order n + 1 matrix

Note

If the argument n is not a positive integer, the function presents an error message and stops.

Author(s)

Frederick Novomestky fnovomes@poly.edu

References

Zhang, Z. and J. Wang (2006). Bernoulli matrix and its algebraic properties, Discrete Applied Nathematics, 154, 1622-1632.

Examples

F <- fibonacci.matrix( 10 )
print( F )

[Package matrixcalc version 1.0-6 Index]