frobenius.matrix {matrixcalc}	R Documentation

Frobenius Matrix

Description

This function returns an order n Frobenius matrix that is useful in numerical mathematics.

Usage

frobenius.matrix(n)

Arguments

n
a positive integer value greater than 1

Details

The Frobenius matrix is also called the companion matrix. It arises in the solution of systems of linear first order differential equations. The formula for the order n Frobenius matrix is {\bf{F}} = ≤ft\lbrack {\begin{array}{ccccc}0&0& \cdots &0&{{{≤ft( { - 1} \right)}^{n - 1}} ≤ft( {\begin{array}{ccccc}n\\0\end{array}} \right)}\\1&0& \cdots &0&{{{≤ft( { - 1} \right)}^{n - 2}} ≤ft( {\begin{array}{ccccc}n\\1\end{array}} \right)}\\0&1& \ddots &0&{{{≤ft( { - 1} \right)}^{n - 3}} ≤ft( {\begin{array}{ccccc}n\\2\end{array}} \right)}\\ \vdots & \vdots & \ddots & \vdots & \vdots \\0&0& \cdots &1&{{{≤ft( { - 1} \right)}^0} ≤ft( {\begin{array}{ccccc}n\\{n - 1}\end{array}} \right)}\end{array}} \right\rbrack.

Value

An order n matrix

Note

If the argument n is not a positive integer that is greater than 1, the function presents an error message and stops.

Author(s)

Frederick Novomestky fnovomes@poly.edu

References

Aceto, L. and D. Trigiante (2001). Matrices of Pascal and Other Greats, American Mathematical Monthly, March 2001, 108(3), 232-245.

Examples

F <- frobenius.matrix( 10 )
print( F )

[Package matrixcalc version 1.0-6 Index]