hankel.matrix {matrixcalc}	R Documentation

Hankel Matrix

Description

This function constructs an order n Hankel matrix from the values in the order n vector x. Each row of the matrix is a circular shift of the values in the previous row.

Usage

hankel.matrix(n, x)

Arguments

n
a positive integer value for order of matrix greater than 1

x
a vector of values used to construct the matrix

Details

A Hankel matrix is a square matrix with constant skew diagonals. The determinant of a Hankel matrix is called a catalecticant. Hankel matrices are formed when the hidden Mark model is sought from a given sequence of data.

Value

An order n matrix.

Note

If the argument n is not a positive integer, the function presents an error message and stops. If the length of x is less than n, the function presents an error message and stops.

Author(s)

Frederick Novomestky fnovomes@poly.edu

References

Power, S. C. (1982). Hankel Operators on Hilbert Spaces, Research notes in mathematics, Series 64, Pitman Publishing.

Examples

H <- hankel.matrix( 4, seq( 1, 7 ) )
print( H )

[Package matrixcalc version 1.0-6 Index]