hilbert.matrix {matrixcalc}	R Documentation

Hilbert matrices

Description

This function returns an n by n Hilbert matrix.

Usage

hilbert.matrix(n)

Arguments

n
Order of the Hilbert matrix

Details

A Hilbert matrix is an order n square matrix of unit fractions with elements defined as H_{i,j} = {1 \mathord{≤ft/ {\vphantom {1 {≤ft( {i + j - 1} \right)}}} \right. } {≤ft( {i + j - 1} \right)}}.

Value

A matrix.

Note

If the argument is less than or equal to zero, the function displays an error message and stops. If the argument is not an integer, the function displays an error message and stops.

Author(s)

Frederick Novomestky fnovomes@poly.edu

References

Hilbert, David (1894). Ein Beitrag zur Theorie des Legendre schen Polynoms, Acta Mathematica, Springer, Netherlands, 18, 155-159.

Examples

H <- hilbert.matrix( 4 )
print( H )

[Package matrixcalc version 1.0-6 Index]