is.non.singular.matrix {matrixcalc}	R Documentation

Test if matrix is non-singular

Description

This function returns TRUE is the matrix argument is non-singular and FALSE otherwise.

Usage

is.non.singular.matrix(x, tol = 1e-08)

Arguments

x
a numeric square matrix

tol
a numeric tolerance level usually left out

Details

The determinant of the matrix x is first computed. If the absolute value of the determinant is greater than or equal to the given tolerance level, then a TRUE value is returned. Otherwise, a FALSE value is returned.

Value

TRUE or FALSE value.

Author(s)

Frederick Novomestky fnovomes@poly.edu

References

Bellman, R. (1987). Matrix Analysis, Second edition, Classics in Applied Mathematics, Society for Industrial and Applied Mathematics.

Horn, R. A. and C. R. Johnson (1990). Matrix Analysis, Cambridge University Press.

See Also

is.singular.matrix

Examples

A <- diag( 1, 3 )
is.non.singular.matrix( A )
B <- matrix( c( 0, 0, 3, 4 ), nrow=2, byrow=TRUE )
is.non.singular.matrix( B )

[Package matrixcalc version 1.0-6 Index]