is.singular.matrix {matrixcalc}	R Documentation

Test for singular square matrix

Description

This function returns TRUE is the matrix argument is singular and FALSE otherwise.

Usage

is.singular.matrix(x, tol = 1e-08)

Arguments

x
a numeric square matrix

tol
a numeric tolerance level usually left out

Details

The determinant of the matrix x is first computed. If the absolute value of the determinant is less than the given tolerance level, then a TRUE value is returned. Otherwise, a FALSE value is returned.

Value

A TRUE or FALSE value.

Author(s)

Frederick Novomestky fnovomes@poly.edu

References

Bellman, R. (1987). Matrix Analysis, Second edition, Classics in Applied Mathematics, Society for Industrial and Applied Mathematics.

Horn, R. A. and C. R. Johnson (1990). Matrix Analysis, Cambridge University Press.

See Also

is.non.singular.matrix

Examples

A <- diag( 1, 3 )
is.singular.matrix( A )
B <- matrix( c( 0, 0, 3, 4 ), nrow=2, byrow=TRUE )
is.singular.matrix( B )

[Package matrixcalc version 1.0-6 Index]