is.skew.symmetric.matrix {matrixcalc}	R Documentation

Test for a skew-symmetric matrix

Description

This function returns TRUE if the matrix argument x is a skew symmetric matrix, i.e., the transpose of the matrix is the negative of the matrix. Otherwise, FALSE is returned.

Usage

is.skew.symmetric.matrix(x, tol = 1e-08)

Arguments

x
a numeric square matrix

tol
a numeric tolerance level usually left out

Details

Let {\bf{x}} be an order n matrix. If every element of the matrix {\bf{x}} + {\bf{x'}} in absolute value is less than the given tolerance, then the matrix argument is declared to be skew symmetric.

Value

A TRUE or FALSE value.

Author(s)

Frederick Novomestky fnovomes@poly.edu

References

Bellman, R. (1987). Matrix Analysis, Second edition, Classics in Applied Mathematics, Society for Industrial and Applied Mathematics.

Horn, R. A. and C. R. Johnson (1990). Matrix Analysis, Cambridge University Press.

Examples

A <- diag( 1, 3 )
is.skew.symmetric.matrix( A )
B <- matrix( c( 0, -2, -1, -2, 0, -4, 1, 4, 0 ), nrow=3, byrow=TRUE )
is.skew.symmetric.matrix( B )
C <- matrix( c( 0, 2, 1, 2, 0, 4, 1, 4, 0 ), nrow=3, byrow=TRUE )
is.skew.symmetric.matrix( C )

[Package matrixcalc version 1.0-6 Index]