lower.triangle {matrixcalc}	R Documentation

Lower triangle portion of a matrix

Description

Returns the lower triangle including the diagonal of a square numeric matrix.

Usage

lower.triangle(x)

Arguments

x
a matrix

Value

A matrix.

Author(s)

Frederick Novomestky fnovomes@poly.edu

References

Bellman, R. (1987). Matrix Analysis, Second edition, Classics in Applied Mathematics, Society for Industrial and Applied Mathematics.

See Also

is.square.matrix

Examples

B <- matrix( seq( 1, 16, 1 ), nrow=4, byrow=TRUE )
lower.triangle( B )

[Package matrixcalc version 1.0-6 Index]