matrix.trace {matrixcalc}	R Documentation

The trace of a matrix

Description

This function returns the trace of a given square numeric matrix.

Usage

matrix.trace(x)

Arguments

x
a matrix

Value

A numeric value which is the sum of the values on the diagonal.

Note

If the argument x is not numeric, the function presents and error message and terminates. If the argument x is not a square matrix, the function presents an error message and terminates.

Author(s)

Frederick Novomestky fnovomes@poly.edu

References

Bellman, R. (1987). Matrix Analysis, Second edition, Classics in Applied Mathematics, Society for Industrial and Applied Mathematics.

Examples

A <- matrix( seq( 1, 16, 1 ), nrow=4, byrow=TRUE )
matrix.trace( A )

[Package matrixcalc version 1.0-6 Index]