maximum.norm {matrixcalc}	R Documentation

Maximum norm of matrix

Description

This function returns the max norm of a real matrix.

Usage

maximum.norm(x)

Arguments

x
a numeric matrix or vector

Details

Let {\bf{x}} be an m \times n real matrix. The max norm returned is ≤ft\| {\bf{x}} \right\|_{\max } = \mathop {\max }\limits_{i,j} ≤ft| {x_{i,j} } \right|.

Value

A numeric value.

Author(s)

Frederick Novomestky fnovomes@poly.edu

References

Bellman, R. (1987). Matrix Analysis, Second edition, Classics in Applied Mathematics, Society for Industrial and Applied Mathematics.

Golub, G. H. and C. F. Van Loan (1996). Matrix Computations, Third Edition, The John Hopkins University Press.

Horn, R. A. and C. R. Johnson (1985). Matrix Analysis, Cambridge University Press.

See Also

inf.norm, one.norm

Examples

A <- matrix( c( 3, 5, 7, 2, 6, 4, 0, 2, 8 ), nrow=3, ncol=3, byrow=TRUE )
maximum.norm( A )

[Package matrixcalc version 1.0-6 Index]