pascal.matrix {matrixcalc}	R Documentation

Pascal matrix

Description

This function returns an n by n Pascal matrix.

Usage

pascal.matrix(n)

Arguments

n
Order of the matrix

Details

In mathematics, particularly matrix theory and combinatorics, the Pascal matrix is a lower triangular matrix with binomial coefficients in the rows. It is easily obtained by performing an LU decomposition on the symmetric Pascal matrix of the same order and returning the lower triangular matrix.

Value

An order n matrix.

Note

If the argument n is not a positive integer, the function presents an error message and stops.

Author(s)

Frederick Novomestky fnovomes@poly.edu

References

Aceto, L. and D. Trigiante, (2001). Matrices of Pascal and Other Greats, American Mathematical Monthly, March 2001, 232-245.

Call, G. S. and D. J. Velleman, (1993). Pascal's matrices, American Mathematical Monthly, April 1993, 100, 372-376.

Edelman, A. and G. Strang, (2004). Pascal Matrices, American Mathematical Monthly, 111(3), 361-385.

See Also

lu.decomposition, symmetric.pascal.matrix

Examples

P <- pascal.matrix( 4 )
print( P )

[Package matrixcalc version 1.0-6 Index]