shift.down {matrixcalc}	R Documentation

Shift matrix m rows down

Description

This function returns a matrix that has had its rows shifted downwards filling the above rows with the given fill value.

Usage

shift.down(A, rows = 1, fill = 0)

Arguments

A
a matrix

rows
the number of rows to be shifted

fill
the fill value which as a default is zero

Value

A matrix.

Note

If the argument A is not a numeric matrix, then the function presents an error message and stops. If the argument rows is not a positive integer, then the function presents an error message and stops.

Author(s)

Frederick Novomestky fnovomes@poly.edu

Examples

A <- matrix( seq( 1, 16, 1 ), nrow=4, byrow=TRUE )
shift.down( A, 1 )
shift.down( A, 3 )

[Package matrixcalc version 1.0-6 Index]