shift.right {matrixcalc}	R Documentation

Shift matrix n columns to the right

Description

This function returns a matrix that has been shifted to the right n columns filling the previous columns with the given fill value.

Usage

shift.right(A, cols = 1, fill = 0)

Arguments

A
a matrix

cols
integer number of columns to be shifted to the right

fill
the fill which as default value zero

Value

A matrix.

Note

If the argument A is not a numeric matrix, then the function presents an error message and stops. If the argument rows is not a positive integer, then the function presents an error message and stops.

Author(s)

Frederick Novomestky fnovomes@poly.edu

Examples

A <- matrix( seq( 1, 16, 1 ), nrow=4, byrow=TRUE )
shift.right( A, 1 )
shift.right( A, 2 )

[Package matrixcalc version 1.0-6 Index]