spectral.norm {matrixcalc}	R Documentation

Spectral norm of matrix

Description

This function returns the spectral norm of a real matrix.

Usage

spectral.norm(x)

Arguments

x
a numeric matrix or vector

Details

Let {\bf{x}} be an m \times n real matrix. The function computes the order n square matrixmatrix {\bf{A}} = {\bf{x'}}\;{\bf{x}}. The R function eigen is applied to this matrix to obtain the vector of eigenvalues {\bf{λ }} = ≤ft\lbrack {\begin{array}{cccc} {λ _1 } & {λ _2 } & \cdots & {λ _n } \\ \end{array}} \right\rbrack. By construction the eigenvalues are in descending order of value so that the largest eigenvalue is λ _1. Then the spectral norm is ≤ft\| {\bf{x}} \right\|_2 = √ {λ _1 }. If {\bf{x}} is a vector, then {\bf{L}}_2 = √ {\bf{A}} is returned.

Value

A numeric value.

Note

If the argument x is not numeric, an error message is displayed and the function terminates. If the argument is neither a matrix nor a vector, an error message is displayed and the function terminates. If the product matrix {\bf{x'}}\;{\bf{x}} is negative definite, an error message displayed and the function terminates.

Author(s)

Frederick Novomestky fnovomes@poly.edu

References

Bellman, R. (1987). Matrix Analysis, Second edition, Classics in Applied Mathematics, Society for Industrial and Applied Mathematics.

Golub, G. H. and C. F. Van Loan (1996). Matrix Computations, Third Edition, The John Hopkins University Press.

Horn, R. A. and C. R. Johnson (1985). Matrix Analysis, Cambridge University Press.

Examples

x <- matrix( c( 2, 4, 2, 1, 3, 1, 5, 2, 1, 2, 3, 3 ), nrow=3, ncol=4, byrow=TRUE )
spectral.norm( x )

[Package matrixcalc version 1.0-6 Index]