symmetric.pascal.matrix {matrixcalc}	R Documentation

Symmetric Pascal matrix

Description

This function returns an n by n symmetric Pascal matrix.

Usage

symmetric.pascal.matrix(n)

Arguments

n
Order of the matrix

Details

In mathematics, particularly matrix theory and combinatorics, the symmetric Pascal matrix is a square matrix from which you can derive binomial coefficients. The matrix is an order n symmetric matrix with typical element given by {S_{i,j}} = {{n!} \mathord{≤ft/ {\vphantom {{n!} {≤ft[ {r!\;≤ft( {n - r} \right)!} \right]}}} \right. } {≤ft[ {r!\;≤ft( {n - r} \right)!} \right]}} where n = i + j - 2 and r = i - 1. The binomial coefficients are elegantly recovered from the symmetric Pascal matrix by performing an LU decomposition as {\bf{S}} = {\bf{L}}\;{\bf{U}}.

Value

An order n matrix.

Note

If the argument n is not a positive integer, the function presents an error message and stops.

Author(s)

Frederick Novomestky fnovomes@poly.edu

References

Call, G. S. and D. J. Velleman, (1993). Pascal's matrices, American Mathematical Monthly, April 1993, 100, 372-376.

Edelman, A. and G. Strang, (2004). Pascal Matrices, American Mathematical Monthly, 111(3), 361-385.

Examples

S <- symmetric.pascal.matrix( 4 )
print( S )

[Package matrixcalc version 1.0-6 Index]