toeplitz.matrix {matrixcalc}	R Documentation

Toeplitz Matrix

Description

This function constructs an order n Toeplitz matrix from the values in the order 2 * n - 1 vector x.

Usage

toeplitz.matrix(n, x)

Arguments

n
a positive integer value for order of matrix greater than 1

x
a vector of values used to construct the matrix

Details

The element T[i,j] in the Toeplitz matrix is x[i-j+n].

Value

An order n matrix.

Note

If the argument n is not a positive integer, the function presents an error message and stops. If the length of x is not equal to 2 * n - 1, the function presents an error message and stops.

Author(s)

Frederick Novomestky fnovomes@poly.edu

References

Monahan, J. F. (2011). Numerical Methods of Statistics, Cambridge University Press.

Examples

T <- toeplitz.matrix( 4, seq( 1, 7 ) )
print( T )

[Package matrixcalc version 1.0-6 Index]