u.vectors {matrixcalc}	R Documentation

u vectors of an identity matrix

Description

This function constructs an order n * ( n + 1 ) / 2 identity matrix and an order matrix u that that maps the ordered pair of indices (i,j) i=j, ..., n; j=1, ..., n to a column in this identity matrix.

Usage

u.vectors(n)

Arguments

n
a positive integer value for the order of underlying matrices

Details

The function firsts constructs an identity matrix of order \frac{1}{2}n≤ft( {n + 1} \right). {{{\bf{u}}_{i,j}}} is the column vector in the order \frac{1}{2}n≤ft( {n + 1} \right) identity matrix for column k = ≤ft( {j - 1} \right)n + i - \frac{1}{2}j≤ft( {j - 1} \right).

Value

A list with two named components

k
order n square matrix that maps each ordered pair (i,j) to a column in the identity matrix

I
order \frac{1}{2}n≤ft( {n + 1} \right) identity matrix

Note

If the argument is not an integer, the function displays an error message and stops. If the argument is less than two, the function displays an error message and stops.

Author(s)

Frederick Novomestky fnovomes@poly.edu

References

Magnus, J. R. and H. Neudecker (1980). The elimination matrix, some lemmas and applications, SIAM Journal on Algebraic Discrete Methods, 1(4), December 1980, 422-449.

Magnus, J. R. and H. Neudecker (1999) Matrix Differential Calculus with Applications in Statistics and Econometrics, Second Edition, John Wiley.

Examples

u <- u.vectors( 3 )

[Package matrixcalc version 1.0-6 Index]