vandermonde.matrix {matrixcalc}	R Documentation

Vandermonde matrix

Description

This function returns an m by n matrix of the powers of the alpha vector

Usage

vandermonde.matrix(alpha, n)

Arguments

alpha
A numerical vector of values

n
The column dimension of the Vandermonde matrix

Details

In linear algebra, a Vandermonde matrix is an m \times n matrix with terms of a geometric progression of an m \times 1 parameter vector {\bf{α }} = {≤ft\lbrack {\begin{array}{cccc} {{α _1}}&{{α _2}}& \cdots &{{α _m}} \end{array}} \right\rbrack^\prime }

such that V≤ft( {\bf{α }} \right) = ≤ft\lbrack {\begin{array}{ccccc} 1&{{α _1}}&{α _1^2}& \cdots &{α _1^{n - 1}}\\ 1&{{α _2}}&{α _2^2}& \cdots &{α _2^{n - 1}}\\ 1&{{α _3}}&{α _3^2}& \cdots &{α _3^{n - 1}}\\ \cdots & \cdots & \cdots & \cdots & \cdots \\ 1&{{α _m}}&{α _m^2}& \cdots &{α _m^{n - 1}} \end{array}} \right\rbrack.

Value

A matrix.

Author(s)

Frederick Novomestky fnovomes@poly.edu

References

Horn, R. A. and C. R. Johnson (1991). Topics in matrix analysis, Cambridge University Press.

Examples

alpha <- c( .1, .2, .3, .4 )
V <- vandermonde.matrix( alpha, 4 )
print( V )

[Package matrixcalc version 1.0-6 Index]