vech {matrixcalc}	R Documentation

Vectorize a matrix

Description

This function returns a stack of the lower triangular matrix of a square matrix as a matrix with 1 column and n * ( n + 1 ) / 2 rows

Usage

vech(x)

Arguments

x
a matrix

Value

A matrix with \frac{1}{2}n≤ft( {n + 1} \right) rows and one column.

Author(s)

Frederick Novomestky fnovomes@poly.edu

References

Magnus, J. R. and H. Neudecker (1999) Matrix Differential Calculus with Applications in Statistics and Econometrics, Second Edition, John Wiley.

See Also

is.square.matrix

Examples

x <- matrix( seq( 1, 16, 1 ), nrow=4, byrow=TRUE )
print( x )
y <- vech( x )
print( y )

[Package matrixcalc version 1.0-6 Index]